+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Bug Fixes
 
 - Fixed exhaust muffler detaching, if the player died when the car was at hom and player was away from it
+- Fixed bucket lid detaching
 
 ## 3.7.2 (08.07.2022)